requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==4.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
inflection==0.3.1
iniconfig==2.0.0
lazy-object-proxy==1.9.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.5
pytest==7.3.1
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-django==4.5.2
pytest-drf==1.1.3
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
wrapt==1.15.0
